requirements.txt
aiohttp
nextcord
motor
thefuzz
pillow
python-dotenv
markdownify
aiosqlite
ImageHash